Warning: Tide Seminar
These are all the files used to create the tide seminar. However, there is direct referencing here to another github directory ShinySlides that I need to fix.
Similarly to my compeletionSeminar there is also hard coding that needs fixing.
I will update both of these things in time. To compile in the meantime remove the custom CSS and comment out the hardcoded pieces.